Revert "Merge remote-tracking branch 'origin/master' into ribdashboard-newstartlineadvantage"
This reverts commit 2124fa8582ba33e2e9e916344cf1c3f1d3e85c38, reversing
changes made to 752906bed4ac1fd1bc03d6c724c32bb41cf27939.

+++ java/com.sap.sailing.server/src/com/sap/sailing/server/simulation/SimulationService.java
@@ -11,7 +11,7 @@ import com.sap.sailing.simulator.SimulationResults;
 
 public interface SimulationService {
 
-    long getSimulationResultsVersion(LegIdentifier legIdentifier);
+    int getSimulationResultsVersion(LegIdentifier legIdentifier);
     
     SimulationResults getSimulationResults(LegIdentifier legIdentifier);
